sim-rs: make message types protocol-specific
